requirements.txt
astroid==2.1.0
cached-property==1.5.1
django-cors-headers===3.0.2
django-extensions==2.1.5
django-filter==2.1.0
django-organizations==1.1.1
django-smart-selects==1.5.4
django-url-filter==0.3.12
Django==2.2.3
djangorestframework-bulk==0.2.1
djangorestframework-jwt==1.11.0
djangorestframework-stubs==0.3.0
djangorestframework==3.10.3
drf-extensions==0.5.0
enum-compat==0.0.2
isort==4.3.4
lazy-object-proxy==1.3.1
Markdown==3.0.1
mccabe==0.6.1
numpy==1.16.2
Pillow==5.4.1
psycopg2-binary==2.7.7
PyJWT==1.7.1
pylint==2.2.2
python-magic==0.4.15
pytz==2018.9
scipy
six==1.12.0
typed-ast==1.3.1
wrapt==1.11.1
